Cleanup shim loading
